Sort out Player to MusicSlider, NextButton, PreviousButton, TogglePlaybackButton
