+++ JangSungwoo1901/MusicApp/App.js
@@ -76,14 +76,14 @@ export default function App() {
       }
     }
   }
-
+  
   return (
     <View>
       <Player
          onPrevious={_skipToPrevious}
          onNext={_skipToNext}
          onTogglePlayback={togglePlayback}
-         ></Player>
+         />
     </View>
   );
 };
